pyproject.toml
[project]
name = "ingestors"
version = "3.22.0"
description = "Ingestors extract the contents of mixed unstructured documents into structured (followthemoney) data. "
authors = [
    { name = "Friedrich Lindenberg", email = "friedrich@pudo.org" },
    { name = "OCCRP Data Team", email = "data@occrp.org" },
    { name = "ID.IO", email = "hi@investigativedata.org" },
]
readme = "README.md"
license = "AGPL-3.0"
requires-python = ">=3.11,<4.0"
dependencies = [
    "banal (==1.0.6)",
    "normality (==2.5.0)",
    "pantomime (==0.6.1)",
    "followthemoney (==3.5.9)",
    "followthemoney-store[postgresql] (>=3.1.0,<3.2.0)",
    "servicelayer @ git+https://github.com/investigativedata/servicelayer.git@main",
    "languagecodes (==1.1.1)",
    "countrytagger (==0.1.2)",
    "pyicu (==2.12)",
    "google-cloud-vision (==3.7.2)",
    "tesserocr (==2.7.1)",
    "spacy (==3.6.1)",
    "numpy (<2.0)",
    "fingerprints (==1.2.3)",
    "fasttext (==0.9.3)",
    "pika (==1.3.2)",
    "nomenklatura (==3.15.2)",
    "dbf (==0.99.9)",
    "pymediainfo (==6.1.0)",
    "python-magic (==0.4.27)",
    "rarfile (==4.2)",
    "xlrd (==2.0.1)",
    "openpyxl (==3.1.2)",
    "odfpy (==1.4.1)",
    "faust-cchardet (==2.1.19)",
    "lxml (==5.0.0)",
    "olefile (==0.47)",
    "Pillow (==10.1.0)",
    "vobject (==0.9.6.1)",
    "msglite (==0.30.0)",
    "icalendar (==5.0.12)",
    "cryptography (==41.0.7)",
    "requests[security] (==2.31.0)",
    "pymupdf (==1.21.1)",
    "prometheus-client (==0.17.1)",
    "sentry_sdk (==2.0.1)",
    # servicelayer extras requirements
    "boto3 (>=1.11.9,<2.0.0)",
    "grpcio (>=1.32.0,<2.0.0)",
    "google-cloud-storage (>=1.31.0,<3.0.0)",
    "faster-whisper (==1.1.1)",
    "filelock (==3.17.0)"
]

[project.scripts]
ingestors = "ingestors.cli:cli"

[project.gui-scripts]
ingestors = "ingestors.cli:cli"

[project.entry-points."ingestors"]
ignore = "ingestors.ignore:IgnoreIngestor"
html = "ingestors.documents.html:HTMLIngestor"
xml = "ingestors.documents.xml:XMLIngestor"
plain = "ingestors.documents.plain:PlainTextIngestor"
office = "ingestors.documents.office:DocumentIngestor"
opendoc = "ingestors.documents.opendoc:OpenDocumentIngestor"
ooxml = "ingestors.documents.ooxml:OfficeOpenXMLIngestor"
djvu = "ingestors.documents.djvu:DjVuIngestor"
pdf = "ingestors.documents.pdf:PDFIngestor"
rar = "ingestors.packages.rar:RARIngestor"
zip = "ingestors.packages.zip:ZipIngestor"
tar = "ingestors.packages.tar:TarIngestor"
7z = "ingestors.packages:SevenZipIngestor"
gz = "ingestors.packages:GzipIngestor"
bz2 = "ingestors.packages:BZ2Ingestor"
pst = "ingestors.email.outlookpst:OutlookPSTIngestor"
olm = "ingestors.email.olm:OutlookOLMArchiveIngestor"
opfmsg = "ingestors.email.olm:OutlookOLMMessageIngestor"
olemsg = "ingestors.email.outlookmsg:OutlookMsgIngestor"
msg = "ingestors.email.msg:RFC822Ingestor"
emlx = "ingestors.email.emlx:AppleEmlxIngestor"
vcard = "ingestors.email.vcard:VCardIngestor"
calendar = "ingestors.email.calendar:CalendarIngestor"
csv = "ingestors.tabular.csv:CSVIngestor"
access = "ingestors.tabular.access:AccessIngestor"
sqlite = "ingestors.tabular.sqlite:SQLiteIngestor"
xls = "ingestors.tabular.xls:ExcelIngestor"
xlsx = "ingestors.tabular.xlsx:ExcelXMLIngestor"
ods = "ingestors.tabular.ods:OpenOfficeSpreadsheetIngestor"
mbox = "ingestors.email.mbox:MboxFileIngestor"
dbf = "ingestors.tabular.dbf:DBFIngestor"
image = "ingestors.media.image:ImageIngestor"
tiff = "ingestors.media.tiff:TIFFIngestor"
svg = "ingestors.media.svg:SVGIngestor"
audio = "ingestors.media.audio:AudioIngestor"
video = "ingestors.media.video:VideoIngestor"
json = "ingestors.misc.jsonfile:JSONIngestor"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "8.2.0"
pytest-cov = "5.0.0"
click = "8.1.7"
